Add files via upload
task_freshworks.py is the main code file which is having all the required functions. Also, it has an update function.
testapplication.py is the command-line statements and working. By running it on terminal/command prompt you will be able to access the functions of task_freshworks.
data.txt will contain all the JSON data.